+++ src/schedule/schedule.module.ts
@@ -5,10 +5,12 @@ import { RecurrenceGroup } from './recurrence-group.entity';
 import { ScheduleService } from './service/schedule.service';
 import { ScheduleWatchService } from './service/schedule-watch.service';
 import { ScheduleRecurringService } from './service/schedule-recurring.service';
-import { ScheduleController } from './schedule.controller';
+import { ScheduleController } from './controller/schedule.controller';
+import { ScheduleRecurringController } from './controller/schedule-recurring.controller';
+import { ScheduleClassRepController } from './controller/schedule-class-rep.controller';
+import { ScheduleWatchController } from './controller/schedule-watch.controller';
 import { ScheduleCronService } from './service/schedule-cron.service';
 import { ScheduleNotificationService } from './service/schedule-notification.service';
-import { ScheduleWatchController } from './schedule-watch.controller';
 import { UserModule } from '../user/user.module';
 import { TagModule } from '../tag/tag.module';
 import { Tag } from '../tag/tag.entity';
@@ -25,7 +27,12 @@ import { GoogleModule } from '../google/google.module';
     ResourceModule,
     GoogleModule,
   ],
-  controllers: [ScheduleController, ScheduleWatchController],
+  controllers: [
+    ScheduleController,
+    ScheduleRecurringController,
+    ScheduleClassRepController,
+    ScheduleWatchController,
+  ],
   providers: [
     ScheduleService,
     ScheduleWatchService,
